backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
requests
python-multipart
neo4j
langchain
langchain-community
langchain-openai
langchain-experimental
httpx
mysql-connector-python
sqlalchemy
Pillow
qrcode
